Created components from the methods of App.js
